more sane type grouping for TypedBandDict
